Add missing bzlib function
In updating bzlib I missed adding an error function required by the newer
version: +bz_internal_error+. This broke the C build, although it had no effect
on the ruby bindings. Fixed!

+++ c/src/scanner_mb.c
@@ -1,3 +1,4 @@
+
 #line 1 "src/scanner_mb.rl"
 /* scanner.rl -*-C-*- */
 #include <ctype.h>
@@ -17,11 +18,12 @@
     goto ret;         \
 } while(0)
 
+
 #line 31 "src/scanner_mb.rl"
 
 
 
-#line 25 "src/scanner_mb.c"
+#line 27 "src/scanner_mb.c"
 static const char _StdTok_actions[] = {
 	0, 1, 0, 1, 1, 1, 2, 1, 
 	3, 1, 4, 1, 5, 1, 17, 1, 
@@ -6218,6 +6220,7 @@ static const int StdTok_error = 0;
 static const int StdTok_en_frt_tokenizer = 37;
 static const int StdTok_en_main = 1;
 
+
 #line 34 "src/scanner_mb.rl"
 
 static const char *position_in_mb( const unsigned int *orig_wc,
@@ -6343,7 +6346,7 @@ void frt_std_scan_mb(const char *in_mb,
     unsigned int *ts = 0, *te = 0;
 
     
-#line 6347 "src/scanner_mb.c"
+#line 6350 "src/scanner_mb.c"
 	{
 	cs = StdTok_start;
 	top = 0;
@@ -6351,6 +6354,7 @@ void frt_std_scan_mb(const char *in_mb,
 	te = 0;
 	act = 0;
 	}
+
 #line 158 "src/scanner_mb.rl"
 
     unsigned int in_wc[4096] = {0};
@@ -6366,7 +6370,7 @@ void frt_std_scan_mb(const char *in_mb,
     *token_size = 0;
 
     
-#line 6370 "src/scanner_mb.c"
+#line 6374 "src/scanner_mb.c"
 	{
 	int _klen;
 	unsigned int _trans;
@@ -6384,10 +6388,10 @@ _resume:
 	while ( _nacts-- > 0 ) {
 		switch ( *_acts++ ) {
 	case 4:
-#line 1 "src/scanner_mb.rl"
+#line 1 "NONE"
 	{ts = p;}
 	break;
-#line 6391 "src/scanner_mb.c"
+#line 6395 "src/scanner_mb.c"
 		}
 	}
 
@@ -6409,7 +6413,7 @@ _resume:
 			else if ( (*p) > *_mid )
 				_lower = _mid + 1;
 			else {
-				_trans += (_mid - _keys);
+				_trans += (unsigned int)(_mid - _keys);
 				goto _match;
 			}
 		}
@@ -6432,7 +6436,7 @@ _resume:
 			else if ( (*p) > _mid[1] )
 				_lower = _mid + 2;
 			else {
-				_trans += ((_mid - _keys)>>1);
+				_trans += (unsigned int)((_mid - _keys)>>1);
 				goto _match;
 			}
 		}
@@ -6454,11 +6458,11 @@ _eof_trans:
 		switch ( *_acts++ )
 		{
 	case 0:
-#line 14 "src/scanner_mb.rl"
+#line 14 "src/url.rl"
 	{ skip = p - ts; }
 	break;
 	case 1:
-#line 26 "src/scanner_mb.rl"
+#line 26 "src/url.rl"
 	{ trunc = 1; }
 	break;
 	case 2:
@@ -6466,139 +6470,139 @@ _eof_trans:
 	{ p--; {stack[top++] = cs; cs = 37; goto _again;} }
 	break;
 	case 5:
-#line 1 "src/scanner_mb.rl"
+#line 1 "NONE"
 	{te = p+1;}
 	break;
 	case 6:
-#line 15 "src/scanner_mb.rl"
+#line 15 "src/scanner.in"
 	{act = 2;}
 	break;
 	case 7:
-#line 16 "src/scanner_mb.rl"
+#line 16 "src/scanner.in"
 	{act = 3;}
 	break;
 	case 8:
-#line 17 "src/scanner_mb.rl"
+#line 17 "src/scanner.in"
 	{act = 4;}
 	break;
 	case 9:
-#line 20 "src/scanner_mb.rl"
+#line 20 "src/scanner.in"
 	{act = 5;}
 	break;
 	case 10:
-#line 23 "src/scanner_mb.rl"
+#line 23 "src/scanner.in"
 	{act = 6;}
 	break;
 	case 11:
-#line 26 "src/scanner_mb.rl"
+#line 26 "src/scanner.in"
 	{act = 7;}
 	break;
 	case 12:
-#line 29 "src/scanner_mb.rl"
+#line 29 "src/scanner.in"
 	{act = 8;}
 	break;
 	case 13:
-#line 32 "src/scanner_mb.rl"
+#line 32 "src/scanner.in"
 	{act = 9;}
 	break;
 	case 14:
-#line 35 "src/scanner_mb.rl"
+#line 35 "src/scanner.in"
 	{act = 10;}
 	break;
 	case 15:
-#line 36 "src/scanner_mb.rl"
+#line 36 "src/scanner.in"
 	{act = 11;}
 	break;
 	case 16:
-#line 40 "src/scanner_mb.rl"
+#line 40 "src/scanner.in"
 	{act = 13;}
 	break;
 	case 17:
-#line 12 "src/scanner_mb.rl"
+#line 12 "src/scanner.in"
 	{te = p+1;{ RET; }}
 	break;
 	case 18:
-#line 17 "src/scanner_mb.rl"
+#line 17 "src/scanner.in"
 	{te = p+1;{ trunc = 2; RET; }}
 	break;
 	case 19:
-#line 29 "src/scanner_mb.rl"
+#line 29 "src/scanner.in"
 	{te = p+1;{ RET; }}
 	break;
 	case 20:
-#line 39 "src/scanner_mb.rl"
+#line 39 "src/scanner.in"
 	{te = p+1;{ return; }}
 	break;
 	case 21:
-#line 40 "src/scanner_mb.rl"
+#line 40 "src/scanner.in"
 	{te = p+1;}
 	break;
 	case 22:
-#line 12 "src/scanner_mb.rl"
+#line 12 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 23:
-#line 15 "src/scanner_mb.rl"
+#line 15 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 24:
-#line 16 "src/scanner_mb.rl"
+#line 16 "src/scanner.in"
 	{te = p;p--;{ trunc = 1; RET; }}
 	break;
 	case 25:
-#line 23 "src/scanner_mb.rl"
+#line 23 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 26:
-#line 26 "src/scanner_mb.rl"
+#line 26 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 27:
-#line 29 "src/scanner_mb.rl"
+#line 29 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 28:
-#line 32 "src/scanner_mb.rl"
+#line 32 "src/scanner.in"
 	{te = p;p--;{ STRIP('.'); }}
 	break;
 	case 29:
-#line 35 "src/scanner_mb.rl"
+#line 35 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 30:
-#line 36 "src/scanner_mb.rl"
+#line 36 "src/scanner.in"
 	{te = p;p--;{ RET; }}
 	break;
 	case 31:
-#line 40 "src/scanner_mb.rl"
+#line 40 "src/scanner.in"
 	{te = p;p--;}
 	break;
 	case 32:
-#line 12 "src/scanner_mb.rl"
+#line 12 "src/scanner.in"
 	{{p = ((te))-1;}{ RET; }}
 	break;
 	case 33:
-#line 15 "src/scanner_mb.rl"
+#line 15 "src/scanner.in"
 	{{p = ((te))-1;}{ RET; }}
 	break;
 	case 34:
-#line 23 "src/scanner_mb.rl"
+#line 23 "src/scanner.in"
 	{{p = ((te))-1;}{ RET; }}
 	break;
 	case 35:
-#line 29 "src/scanner_mb.rl"
+#line 29 "src/scanner.in"
 	{{p = ((te))-1;}{ RET; }}
 	break;
 	case 36:
-#line 35 "src/scanner_mb.rl"
+#line 35 "src/scanner.in"
 	{{p = ((te))-1;}{ RET; }}
 	break;
 	case 37:
-#line 40 "src/scanner_mb.rl"
+#line 40 "src/scanner.in"
 	{{p = ((te))-1;}}
 	break;
 	case 38:
-#line 1 "src/scanner_mb.rl"
+#line 1 "NONE"
 	{	switch( act ) {
 	case 2:
 	{{p = ((te))-1;} RET; }
@@ -6636,7 +6640,7 @@ _eof_trans:
 	}
 	}
 	break;
-#line 6640 "src/scanner_mb.c"
+#line 6644 "src/scanner_mb.c"
 		}
 	}
 
@@ -6646,10 +6650,10 @@ _again:
 	while ( _nacts-- > 0 ) {
 		switch ( *_acts++ ) {
 	case 3:
-#line 1 "src/scanner_mb.rl"
+#line 1 "NONE"
 	{ts = 0;}
 	break;
-#line 6653 "src/scanner_mb.c"
+#line 6657 "src/scanner_mb.c"
 		}
 	}
 
@@ -6668,6 +6672,7 @@ _again:
 
 	_out: {}
 	}
+
 #line 172 "src/scanner_mb.rl"
 
     if ( cs == StdTok_error )
